requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0

# System & Hardware Monitoring
psutil==5.9.6
paho-mqtt==1.6.1
aiofiles==23.2.1

# Phase 2: Voice Interface
openai-whisper==20231117
gtts==2.4.0
librosa==0.10.0
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
pydub==0.25.1

# Phase 3: Vector Databases
scikit-learn==1.3.2

# Phase 4: Domain Agents (no additional dependencies needed)
# Phase 5: Safety Guardrails (Python stdlib only)
# Phase 6: Observability (Python stdlib only)

# Optional: Audio I/O (may fail on macOS, gracefully degraded)
# pyaudio==0.2.13

# Optional: Monitoring & Metrics Export
prometheus-client==0.18.0
python-json-logger==2.0.7

# CLI Tool
click==8.1.7
rich==13.6.0
tabulate==0.9.0

# Environment & Security
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
